Simplified changes a good bit
