+++ .github/DEVELOPMENT.md
@@ -11,20 +11,6 @@ pnpm install
 > This repository includes a list of suggested VS Code extensions.
 > It's a good idea to use [VS Code](https://code.visualstudio.com) and accept its suggestion to install them, as they'll help with development.
 
-## Building
-
-Run [**tsup**](https://tsup.egoist.dev) locally to build source files from `src/` into output files in `lib/`:
-
-```shell
-pnpm build
-```
-
-Add `--watch` to run the builder in a watch mode that continuously cleans and recreates `lib/` as you save files:
-
-```shell
-pnpm build --watch
-```
-
 ## Formatting
 
 [Prettier](https://prettier.io) is used to format code.
@@ -38,7 +24,7 @@ pnpm format --write
 
 ## Linting
 
-[ESLint](https://eslint.org) is used with with [typescript-eslint](https://typescript-eslint.io)) to lint JavaScript and TypeScript source files.
+[ESLint](https://eslint.org) is used to lint JavaScript source files.
 You can run it locally on the command-line:
 
 ```shell
@@ -52,19 +38,3 @@ pnpm run lint --fix
 ```
 
 Note that you'll likely need to run `pnpm build` before `pnpm lint` so that lint rules which check the file system can pick up on any built files.
-
-## Type Checking
-
-You should be able to see suggestions from [TypeScript](https://typescriptlang.org) in your editor for all open files.
-
-However, it can be useful to run the TypeScript command-line (`tsc`) to type check all files in `src/`:
-
-```shell
-pnpm tsc
-```
-
-Add `--watch` to keep the type checker running in a watch mode that updates the display as you save files:
-
-```shell
-pnpm tsc --watch
-```
