--- .github/workflows/tsc.yml
@@ -1,15 +1,0 @@
-jobs:
-  type_check:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: ./.github/actions/prepare
-      - run: pnpm tsc
-
-name: Type Check
-
-on:
-  pull_request: ~
-  push:
-    branches:
-      - main
