--- .jshintrc
@@ -1,59 +1,0 @@
-{
-	"maxerr": 100,
-
-	"browser": true,
-
-	"node": true,
-	"rhino": false,
-	"couch": false,
-	"wsh": false,
-
-	"jquery": true,
-	"prototypejs": false,
-	"mootools": false,
-	"dojo": false,
-
-	"predef": ["escape", "grunt", "Pilotfish", "QUnit"],
-
-	"debug": false,
-	"devel": false,
-
-	"es5": false,
-	"strict": false,
-	"globalstrict": false,
-
-	"asi": false,
-	"laxbreak": true,
-	"bitwise": true,
-	"boss": true,
-	"curly": true,
-	"eqeqeq": false,
-	"eqnull": false,
-	"evil": false,
-	"expr": true,
-	"forin": false,
-	"immed": true,
-	"latedef": false,
-	"loopfunc": true,
-	"noarg": true,
-	"regexp": true,
-	"regexdash": false,
-	"scripturl": true,
-	"shadow": false,
-	"supernew": false,
-	"undef": true,
-	"unused": true,
-
-	"newcap": false,
-	"noempty": true,
-	"nonew": true,
-	"nomen": false,
-	"onevar": false,
-	"plusplus": false,
-	"sub": false,
-	"trailing": true,
-	"indent": 4,
-	"maxstatements": 20,
-	"maxcomplexity": 8,
-	"maxlen": 160
-}
