+++ .vscode/settings.json
@@ -9,11 +9,8 @@
 		"json",
 		"jsonc",
 		"markdown",
-		"typescript",
-		"typescriptreact",
 		"yaml"
 	],
 	"eslint.rules.customizations": [{ "rule": "*", "severity": "warn" }],
-	"eslint.useFlatConfig": true,
-	"typescript.tsdk": "node_modules/typescript/lib"
+	"eslint.useFlatConfig": true
 }
