+++ base64.js
@@ -4,6 +4,7 @@
  *  http://www.webtoolkit.info/
  *
  **/
+// eslint-disable-next-line no-unused-vars
 const Base64 = {
 	// private property
 	_keyStr: "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=",
