--- test/qunit-pjv.html
@@ -1,15 +1,0 @@
-<!doctype html>
-<html>
-	<head>
-		<meta charset="utf-8" />
-		<title>QUnit Tests for package-json-validator</title>
-		<link rel="stylesheet" href="lib/qunit.css" />
-		<script src="lib/qunit.js"></script>
-		<script src="../PJV.js"></script>
-	</head>
-	<body>
-		<div id="qunit"></div>
-
-		<script src="qunit-pjv.js"></script>
-	</body>
-</html>
